# Conflicts:
#	src/main/java/app/domain/model/CSVFileReader.java

+++ src/main/java/app/adapter/SortByTin.java
@@ -16,7 +16,7 @@ public class SortByTin implements SortingAlgorithms {
     public List<ClientDTO> orderClientList (List<ClientDTO> list) {
         for (int i = 0;i<list.size();i++) {
             for (int j=0;j< list.size();j++){
-                if (Integer.parseInt(list.get(i).getTin()) < Integer.parseInt(list.get(j).getTin()) ){
+                if (Double.parseDouble(list.get(i).getTin()) < Double.parseDouble(list.get(j).getTin()) ){
                         ClientDTO temp = list.get(j);
                         list.set(j,list.get(i));
                         list.set(i,temp);
@@ -25,4 +25,5 @@ public class SortByTin implements SortingAlgorithms {
         }
         return list;
     }
+
 }
